Changed some "I" to "we" and updated "Getting started".
I'm working to replace all the "I" language with "we" language, and this change includes some of that. I also updated some the the content in the "Getting started" section so it talks about forking this project. I'm also not sure how we want to handle listing groups in the new Github land. Have them issue pull requests back onto the class fork of this starter code?