Remove references to `array_merge` and `mergesort` from the writeup.
This removes all references to those two parts (since they'll be in the next lab). I also did an editing pass through the rest of the write-up. I _haven't_, however, done an editing pass on the two actual problems yet, and I still need to move the credits info to its own file.